refactor(operations): split pure resolvers out of call.ts
src/operations/call.ts sat at 628 lines against a 600-line limit,
grandfathered at exactly its own size. The ratchet therefore forbade ANY
growth, which blocked two unrelated pieces of work that each needed a
single line in the hopCtx literal:

  - threading contextToolRunCounter so pull.maxCallsPerRun is a real
    per-run cap rather than resetting every hop (noted in #1455)
  - pullCalls telemetry, spec'd at SPEC-context-engine-v2.md:245 and
    AC-18 (noted in #1458)

Extracts the pure helper block ahead of callOp into
src/operations/call-resolvers.ts: newCorrelationId, normalizeRunOutcome,
normalizeSelector, resolveOpModel, resolveTimeoutMs, resolveOpRetry,
synthesizeStory, and MAX_COMPLETE_RETRY_ATTEMPTS. They normalise an
Operation's declarative fields into concrete values, hold no state, touch
no I/O, and never reference callOp's internals -- which is why they can
live apart from the ~450-line dispatch function.

call.ts 628 -> 550 lines, now UNDER the limit rather than grandfathered;
the file-size baseline drops 13 -> 12 entries and call.ts is removed from
it entirely, so it can never silently creep back over.

Behaviour is unchanged: this is a move, not a rewrite. External consumers
are unaffected because they all import through src/operations/index.ts,
which now re-exports newCorrelationId from the new module.

Verified the blocker is actually gone, not merely reduced: adding the
previously-forbidden contextToolRunCounter line to hopCtx puts call.ts at
551 lines and check:file-sizes passes.

FULL suite green (unit + integration + ui), typecheck and lint clean, all
ratchets at baseline.
